redid autocomplete box
ran into a really fascinating issue today, so MUI Autocomplete is really slow when I put in 3000 elements. So I wanted to find a solution for that. I ran into virtualize AC on the MUI site and it's so fast and smooth. However, I ran into another issue, the text would be cut off, I fixed this by removing "noWrap" on the typography element, I then also increased the space between the boxes for readability, everything works well now